pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["mkantonio <mkchoque@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.3"
et-xmlfile = {version = "1.1.0", markers = "python_version >= \"3.6\""}
numpy = {version = "1.24.4", markers = "python_version < \"3.10\""}
openpyxl = "3.1.2"
pandas = "2.0.3"
python-dateutil = {version = "2.8.2", markers = "python_version >= \"2.7\" and python_version not in \"3.0, 3.1, 3.2, 3.3\""}
pytz = "2023.3"
six = {version = "1.16.0", markers = "python_version >= \"2.7\" and python_version not in \"3.0, 3.1, 3.2, 3.3\""}
tzdata = {version = "2023.3", markers = "python_version >= \"2\""}
djangorestframework = "^3.14.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
